tests/wpt/tests/html/semantics/scripting-1/the-script-element/execution-timing/150-noimport.html

<!DOCTYPE html>
<html>
<head>
  <title>Stylesheet in BODY blocking scripts</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<body>
  <div id="test">Test</div>
  <!-- this stylesheet blocks scripts -->
  <link rel="stylesheet" href="css/background.css?pipe=trickle(d2)">
  <script>
    test(function() {
      assert_equals(getComputedStyle(document.getElementById("test")).position,
                    "fixed");
    });
  </script>
</body>
</html>
